Move getSwapInfo action
